refactor(FON-468-1): migrate summary to feature-first architecture
