+++ apps/client/src/features/summary/SummaryNotFound.tsx
@@ -5,10 +5,9 @@ import TechnicalError from '@codegouvfr/react-dsfr/picto/TechnicalError';
 import { useContext } from 'react';
 
 import { OvoidBackground, OvoidMotif } from '@/components/shared/ovoid';
+import { SummaryContext } from '@/features/summary/context/SummaryContext';
 import { useIsSg } from '@/hooks/roles.hook';
 import { ROUTE_PATHS } from '@/utils/route-path.utils';
-
-import { SummaryContext } from './SummaryContext';
 
 /** @see https://www.systeme-de-design.gouv.fr/version-courante/fr/modeles/pages-types/page-d-erreurs */
 export function SummaryNotFound() {
