+++ apps/client/src/features/summary/components/SummaryAutoSaveNotice.tsx
@@ -1,7 +1,7 @@
 import Notice from '@codegouvfr/react-dsfr/Notice';
 
+import { useSummary } from '@/features/summary/context/SummaryContext';
 import { useLocallyStoredState } from '@/hooks/useLocallyStoredState.hook';
-import { useSummary } from '@/pages/summary/SummaryContext';
 
 export function SummaryAutoSaveAlert() {
   const { canWriteSummary } = useSummary();
